feat: add technician repository and scheduling use case
