+++ backend/src/application/repair/get-available-technicians.use-case.ts
@@ -1,9 +1,9 @@
 import { Injectable, Inject } from "@nestjs/common";
-import { REPAIR_REPOSITORY, RepairRepositoryPort } from "../../domain/repair/repair.port";
 import {
-  Technician,
-  TechnicianSpecialty,
-} from "../../domain/repair/repair.entity";
+  TECHNICIAN_REPOSITORY,
+  TechnicianRepositoryPort,
+} from "../../domain/repair/technician.repository.port";
+import { Technician, TechnicianSpecialty } from "../../domain/repair/repair.entity";
 
 export interface GetAvailableTechniciansQuery {
   specialty: TechnicianSpecialty;
@@ -14,8 +14,8 @@ export interface GetAvailableTechniciansQuery {
 @Injectable()
 export class GetAvailableTechniciansUseCase {
   constructor(
-    @Inject(REPAIR_REPOSITORY)
-    private readonly repairRepository: RepairRepositoryPort
+    @Inject(TECHNICIAN_REPOSITORY)
+    private readonly technicianRepository: TechnicianRepositoryPort
   ) {}
 
   async execute(query: GetAvailableTechniciansQuery): Promise<Technician[]> {
@@ -28,12 +28,12 @@ export class GetAvailableTechniciansUseCase {
       throw new Error("Region and city are required");
     }
 
-    return await this.repairRepository.getAvailableTechnicians(
-      query.specialty,
-      {
+    return await this.technicianRepository.findAvailable({
+      specialty: query.specialty,
+      location: {
         region: query.region,
         city: query.city,
       },
-    );
+    });
   }
 }
