+++ backend/src/domain/repair/repair.port.ts
@@ -1,9 +1,7 @@
 import {
   RepairRequest,
-  Technician,
   RepairAppointment,
   RepairEstimate,
-  TechnicianSpecialty,
   AppointmentTimeSlot,
 } from "./repair.entity";
 
@@ -13,7 +11,6 @@ import {
  * (Prisma, Supabase, InMemory, etc.) à leurs ports de domaine.
  */
 export const REPAIR_REPOSITORY = Symbol("RepairRepositoryPort");
-export const REPAIR_DOMAIN_SERVICE = Symbol("RepairDomainServicePort");
 
 /* ------------------------------------------------------------------
  * Repository Port
@@ -31,14 +28,6 @@ export interface RepairRepositoryPort {
   ): Promise<RepairRequest>;
   deleteRepairRequest(id: string): Promise<void>;
 
-  // Technicians
-  getTechnicianById(id: string): Promise<Technician | null>;
-  getAvailableTechnicians(
-    specialty: TechnicianSpecialty,
-    location: { region: string; city: string },
-  ): Promise<Technician[]>;
-  updateTechnicianAvailability(id: string, isAvailable: boolean): Promise<void>;
-
   // Appointments
   createAppointment(
     appointment: Omit<RepairAppointment, "id" | "createdAt" | "updatedAt">,
@@ -54,11 +43,6 @@ export interface RepairRepositoryPort {
     updates: Partial<RepairAppointment>,
   ): Promise<RepairAppointment>;
   cancelAppointment(id: string): Promise<void>;
-  checkTimeSlotAvailability(
-    technicianId: string,
-    date: Date,
-    timeSlot: AppointmentTimeSlot,
-  ): Promise<boolean>;
 
   // Estimates
   createEstimate(
@@ -68,26 +52,4 @@ export interface RepairRepositoryPort {
     repairRequestId: string,
   ): Promise<RepairEstimate[]>;
   getEstimateById(id: string): Promise<RepairEstimate | null>;
-}
-
-/* ------------------------------------------------------------------
- * Domain Service Port
- * ------------------------------------------------------------------ */
-export interface RepairDomainServicePort {
-  calculateRepairCost(
-    deviceType: string,
-    issueType: string,
-  ): Promise<{ min: number; max: number }>;
-  findBestTechnician(
-    specialty: TechnicianSpecialty,
-    location: { region: string; city: string },
-  ): Promise<Technician | null>;
-  getAvailableTimeSlots(
-    technicianId: string,
-    date: Date,
-  ): Promise<AppointmentTimeSlot[]>;
-  sendAppointmentNotification(
-    appointmentId: string,
-    type: "confirmation" | "reminder" | "cancellation",
-  ): Promise<void>;
 }
